feat(useAssignments): separate fetch, filter, and structure
- Moved fetch logic to `fetch.ts` to handle assignment fetching
- Added `filter.ts` to filter assignments based on statuses
- Created `structure.ts` to categorize assignments into time slots
- Updated `useAssignments.tsx` to use the new modular functions
